The circle timer is "drawn on" clockwise
The circle timer is "drawn on" clockwise instead of "erased". When the timer reaches the end it also freezes in it's fully drawn state instead of automatically disappearing/reseting itself. This is the first time i have ever attempted to "fork" something. I pasted the updated implementation file here with changes to lines 134 and 146-175.